Revert "Move jobs into their own module"
This reverts commit b3ef2e48a783e9d66be20aec7c80ca3d76322e69.

Conflicts:
	src/job.rs
